REF rename Application to ConsoleAppl  and Menu to ConsoleMenu. Change placement to menu package
