+++ src/main/java/com/javarush/siberia/StartCryptoApp.java
@@ -1,7 +1,8 @@
 package com.javarush.siberia;
 
 import com.javarush.siberia.constants.Constants;
-import com.javarush.siberia.menu.Menu;
+import com.javarush.siberia.menu.ConsoleApplication;
+import com.javarush.siberia.menu.ConsoleMenu;
 
 public class StartCryptoApp {
     public static void main(String[] args) {
@@ -9,8 +10,8 @@ public class StartCryptoApp {
         System.out.println("Добро пожаловать в примитивный шифровальщик, который Цезарь завещал.");
         System.out.println(Constants.ALPHABET_LENGTH_TXT);
 
-        Application application = new Application();
-        Menu menu = new Menu(application);
-        menu.showMenu();
+        ConsoleApplication consoleApplication = new ConsoleApplication();
+        ConsoleMenu consoleMenu = new ConsoleMenu(consoleApplication);
+        consoleMenu.showMenu();
     }
 }
